[✨ Feat]user signup
유저 회원가입

+++ user/models.py
@@ -61,7 +61,7 @@ class CustomUser(AbstractBaseUser, PermissionsMixin):
 
 
 class Agreements(models.Model):
-    id = models.BigIntegerField(primary_key=True)
+    id = models.AutoField(primary_key=True)  # BigIntegerField에서 AutoField로 변경
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     terms_url = models.CharField(max_length=255)
     agreed_at = models.DateTimeField()
